@@ -6,7 +6,7 @@
   "keywords": [],
   "window": {
     "icon": "app/images/icon.png",
-    "toolbar": false,
+    "toolbar": true,
     "frame": true,
     "width": 800,
     "height": 600,
@@ -14,32 +14,46 @@
     "resizable": true
   },
   "webkit": {
-    "plugin": false
+    "plugin": true
   },
   "main": "app/index.html",
   "dependencies": {
-    "react": "^0.13.3"
+    "react": "^0.14.0"
   },
   "devDependencies": {
     "babelify": "^6.3.0",
     "browserify": "^11.1.0",
     "chai": "^3.2.0",
+    "imap": "^0.8.15",
     "mocha": "^2.0.1",
+    "nodemailer": "^1.8.0",
     "nw-builder": "^2.0.2",
+    "react": "^0.14.0",
+    "react-bootstrap": "^0.27.1",
+    "react-dom": "^0.14.0",
+    "react-redux": "^3.1.0",
     "reactify": "^1.1.1",
+    "redux": "^3.0.2",
+    "redux-logger": "^2.0.3",
+    "redux-thunk": "^1.0.0",
     "watchify": "^3.4.0"
   },
   "scripts": {
     "start": "watchify app/jsx/index.jsx -o app/bundle.js -v",
     "test": "mocha",
-    "app": "nwbuild --run .",
+    "app": "nwbuild --v 0.12.3 --run .",
     "build": "./build/build",
     "clean": "rm -rf build/cache app/bundle.js"
   },
   "browserify": {
     "transform": [
       [
-        "babelify"
+        "babelify",
+        {
+          "optional": [
+            "es7"
+          ]
+        }
       ]
     ]
   }
