@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-module.exports = React.createClass({
+export default React.createClass({
 
   render: function() {
     return (
