@@ -1,6 +1,13 @@
 /* @jsx React.DOM */
 
 var React = require('react');
-var HelloWorld = require('./components/HelloWorld.jsx');
+var ReactDOM = require('react-dom');
+var App = require('./containers/App.jsx');
+import configureStore from '../stores/stores/configStore.js';
+import { Provider } from 'react-redux';
 
-React.render(<HelloWorld />, document.getElementById('content'));
+const store = configureStore();
+
+ReactDOM.render(<Provider store={store}>
+    <App />
+</Provider>, document.getElementById('content'));
